# backend/swagger/noql-api.yaml
openapi: 3.0.3
info:
  title: NoQL API
  version: 1.0.0
  description: NoQL API
  contact:
    name: NoQL Team
    email: babakhonza@gmail.com
servers:
  - url: http://localhost:8080
    description: Development server
...
security:
  - BearerAuth: [ ]
paths:
...
  /chat:
    post:
      tags:
        - Chat
      security:
        - BearerAuth: [ ]
      summary: Create new chat
      description: >
        Create a new chat associated with specified database. Default chat name is "New chat".
        Role USER can create only chats associated with his own databases. Role ADMIN can create chats associated with
        any database.
      parameters:
        - $ref: '#/components/parameters/databaseIdPath'
      responses:
        '201':
          description: Created chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
              example:
                id: "98f2f987-da0e-4b37-afc5-839bd0e5b803"
                name: "New chat"
                messages: [ ]
                modificationDate: "2024-05-24T18:40:24.649+00:00"
        '401':
          description: Unauthorized - invalid or expired access token
        '403':
          description: Forbidden - user is not an ADMIN or owner of the database to which the chat is associated
        '404':
          description: Database of specified ID not found
  /chat/{chatId}:
    get:
      tags:
        - Chat
      security:
        - BearerAuth: [ ]
      summary: Get a chat by ID
      description: Role USER can request only his own chats. Role ADMIN can request any chat.
      parameters:
        - $ref: '#/components/parameters/chatIdPath'
      responses:
        '200':
          description: A chat of specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '401':
          description: Unauthorized - invalid or expired access token
        '403':
          description: Forbidden - user is not an ADMIN or owner of database to which the chat is associated
        '404':
          description: Chat of specified ID not found
    delete:
      tags:
        - Chat
      security:
        - BearerAuth: [ ]
      summary: Delete chat by ID
      description: Role USER can delete only his own chats. Role ADMIN can delete any chat.
      parameters:
        - $ref: '#/components/parameters/chatIdPath'
      responses:
        '204':
          description: No content
        '401':
          description: Unauthorized - invalid or expired access token
        '403':
          description: Forbidden - user is not an ADMIN or owner of database to which the chat is associated
  /chat/{chatId}/name:
    put:
      tags:
        - Chat
      security:
        - BearerAuth: [ ]
      summary: Rename chat by ID
      description: Role USER can rename only his own chats. Role ADMIN can rename any chat.
      parameters:
        - $ref: '#/components/parameters/chatIdPath'
        - name: name
          in: query
          description: New name
          required: true
          schema:
            $ref: '#/components/schemas/ChatName'
      responses:
        '204':
          description: No content
        '401':
          description: Unauthorized - invalid or expired access token
        '403':
          description: Forbidden - user is not an ADMIN or owner of database to which the chat is associated
  /chat/{chatId}/messages:
    post:
      tags:
        - Chat
      security:
        - BearerAuth: [ ]
      summary: Add message to chat.
      description: >
        Add message to chat. Used only for development purposes. Does not verify if the message is valid.
        Role USER can add messages only to his own chats. Role ADMIN can add messages to any chat.
      parameters:
        - $ref: '#/components/parameters/chatIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatQueryWithResponseRequest'
      responses:
        '204':
          description: No content
        '401':
          description: Unauthorized - invalid or expired access token
        '403':
          description: Forbidden - user is not an ADMIN or owner of database to which the chat is associated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "98f2f987-da0e-4b37-afc5-839bd0e5b803"
...
    RetrievedData:
      type: object
      description: Retrieved data from the database query in form of table.
      properties:
        columnNames:
          type: array
          items:
            type: string
          example: [ "sex", "count" ]
        rows:
          type: array
          items:
            type: array
            items:
              type: string
          example: [ [ "male", "8" ], [ "female", "2" ] ]
        page:
          type: integer
          format: int64
          example: 0
        pageSize:
          type: integer
          format: int64
          example: 10
        totalCount:
          type: integer
          format: int64
          example: 200
...
    ChatResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RetrievedData'
        messageId:
          $ref: '#/components/schemas/UUID'
        nlQuery:
          $ref: '#/components/schemas/NaturalLanguageQuery'
        dbQuery:
          $ref: '#/components/schemas/DatabaseLanguageQuery'
        plotUrl:
          type: string
          description: URL to the plot if the users requested a plot
          example: "http://localhost:8080/static/images/13eae4d2-6d7b-41c6-93d0-8f16e4721504-13eae4d2-6d7b-41c6-93d0-8f16e4721555.png"
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        error:
          $ref: '#/components/schemas/QueryErrorMessage'
...
    CreateChatQueryWithResponseRequest:
      type: object
      properties:
        nlQuery:
          $ref: '#/components/schemas/NaturalLanguageQuery'
        llmResult:
          $ref: '#/components/schemas/LLMResult'
        timestamp:
          $ref: '#/components/schemas/Timestamp'
    NaturalLanguageQuery:
      type: string
      description: Natural language query
      example: "Plot sex of users older than 25"
    DatabaseLanguageQuery:
      type: string
      description: Database query language (SQL, ...)
      example: "SELECT * FROM customer;"
    LLMResult:
      type: object
      properties:
        databaseQuery:
          type: string
          description: Query in database specific language (SQL, ...)
          example: "SELECT sex, COUNT(*) FROM public.user WHERE age > 25 GROUP BY sex;"
          nullable: true
        plotUrl:
          type: string
          description: URL to the plot if the users requested a plot
          example: "http://localhost:8080/static/images/13eae4d2-6d7b-41c6-93d0-8f16e4721504.png"
          nullable: true
    QueryErrorMessage:
      type: string
      nullable: true
      example: "Syntax error on position: 30. Table teachers does not exist."
    Timestamp:
      type: string
      format: date-time
      example: "2024-05-24T18:40:24.649+00:00"
...
    Chat:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          $ref: '#/components/schemas/ChatName'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatResponse'
        modificationDate:
          $ref: '#/components/schemas/Timestamp'
        databaseId:
          $ref: '#/components/schemas/UUID'
    ChatName:
      type: string
      example: "Find oldest student"
      minLength: 1
      maxLength: 32
...
  parameters:
...
    databaseIdPath:
      name: databaseId
      in: path
      required: true
      description: ID of the database
      schema:
        $ref: '#/components/schemas/UUID'
...
    chatIdPath:
      name: chatId
      in: path
      required: true
      description: ID of the chat
      schema:
        $ref: '#/components/schemas/UUID'
...